Updateded the scripts to get them all external. removed inventory screen, but left for potential use later as some sort of equipage screen.
Removed default level, replaced with simple level generation. Just
makes a square platform for now. But it is something to go off of.

I need to get some sort of sensor in to the game now to mark when
the player has completed the level. I also should make a block that
can be a visual cue for the player. Perhaps also a compas so that
the player has a general idea of where they are supposed to be
going if the objective is far out of sight.

But wonderful progress never the less. As mysterous as this code
base is, I'm starting to get a basic understanding of it. woohoo!